requirements.txt
Django==4.0.0
pydantic==1.9.1

djangorestframework==3.13.1
pytest==7.1.2
drf_yasg2==1.19.4

fastapi==0.78.0
psycopg2-binary==2.9.3
